pyproject.toml
[tool.poetry]
name = "dis-bot"
version = "0.1.0"
description = ""
authors = ["suisosuii <hayatokomiyama@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
typing-extensions = "^4.9.0"
discord-py = "^2.3.2"
py-cord = "^2.4.1"
flask = "^3.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
